Changed some stuff (everything), added inputs.py.

+++ common.py
@@ -4,4 +4,4 @@ def toPixels((x,y)):
   return((64*x+32,64*y+32))
   
 def toGrid((x,y)):
-  return (floor(x/64),floor(y/64))
+  return (int(floor(x/64)),int(floor(y/64)))
